pyproject.toml
[tool.poetry]
name = "diplom"
version = "0.1.0"
description = ""
authors = ["N777 <Egorov_vadim2001@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
djangorestframework = "^3.14.0"
pandas = "^1.5.2"
lxml = "^4.9.1"
twill = "^3.1"
django-filter = "^22.1"
django-cors-headers = "^3.13.0"
djoser = "^2.1.0"
drf-yasg = "^1.21.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
